% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap2anndata.R
\name{finemap2anndata}
\alias{finemap2anndata}
\title{Convert finemap files to AnnData}
\usage{
finemap2anndata(finemap_files, output_file, panel)
}
\arguments{
\item{finemap_files}{A character vector of file paths to the finemap `.rds` files.}

\item{output_file}{A character string specifying the path to save the resulting AnnData object as an `.h5ad` file.}

\item{panel}{A character string specifying the SNP genotyping/imputation panel}
}
\value{
The AnnData object containing the sparse matrix of lABF values and metadata.
}
\description{
This function reads finemap `.rds` files, filters rows where `is_cs` is `TRUE`,
and creates a sparse matrix of lABF values for SNPs across credible sets.
The function then converts the sparse matrix to an AnnData object and
includes metadata about the start and end positions of each credible set.
}
\examples{
\dontrun{
# Load necessary libraries
library(anndata)
library(data.table)
library(dplyr)
library(Matrix)

# Define the folder containing the finemap files
finemap_folder <- "/group/pirastu/prj_014_huvec_coloc/2024_04_huvec_hcoloc_analysis_V2/results/finemap/"

# List all files matching the pattern
finemap_chr22_files <- list.files(finemap_folder, pattern = "chr22.*\\\\.rds", full.names = TRUE)
finemap_chr22_files <- finemap_chr22_files[!grepl("GWAS", finemap_chr22_files)]

finemap_chr21_files <- list.files(finemap_folder, pattern = "chr21.*\\\\.rds", full.names = TRUE)
finemap_chr21_files <- finemap_chr21_files[!grepl("GWAS", finemap_chr21_files)]

# List all ATAC files matching the pattern and filter out those containing "GWAS"
finemap_chr22_ATAC_files <- list.files(finemap_folder, pattern = "^ATAC.*\\\\chr22.*\\\\.rds", full.names = TRUE)
finemap_chr22_ATAC_files <- finemap_chr22_ATAC_files[!grepl("GWAS", finemap_chr22_ATAC_files)]

# List all ChipSeq files matching the pattern and filter out those containing "GWAS"
finemap_chr22_ChipSeq_files <- list.files(finemap_folder, pattern = "^ChipSeq.*\\\\chr22.*\\\\.rds", full.names = TRUE)
finemap_chr22_ChipSeq_files <- finemap_chr22_ChipSeq_files[!grepl("GWAS", finemap_chr22_ChipSeq_files)]

# Define the output file paths
output_file_chr22 <- file.path("/group/pirastu/prj_013_horizontal_codec/2024_06_20_AnnData", "HUVEC_chr22_combined_credible_sets.h5ad")
output_file_chr21 <- file.path("/group/pirastu/prj_013_horizontal_codec/2024_06_20_AnnData", "HUVEC_chr21_combined_credible_sets.h5ad")
output_file_chr22_ATAC <- file.path("/group/pirastu/prj_013_horizontal_codec/2024_06_20_AnnData", "HUVEC_chr22_ATAC_combined_credible_sets.h5ad")
output_file_chr22_ChipSeq <- file.path("/group/pirastu/prj_013_horizontal_codec/2024_06_20_AnnData", "HUVEC_chr22_ChipSeq_combined_credible_sets.h5ad")

ad_chr_22 <- finemap2anndata(
  finemap_files = finemap_chr22_files,
  output_file = output_file_chr22,
  panel = "HRC"
)

ad_chr_21 <- finemap2anndata(
  finemap_files = finemap_chr21_files,
  output_file = output_file_chr21,
  panel = "HRC"
)

ad_chr_22_ATAC <- finemap2anndata(
  finemap_files = finemap_chr22_ATAC_files,
  output_file = output_file_chr22_ATAC,
  panel = "HRC"
)

ad_chr_22_ChipSeq <- finemap2anndata(
  finemap_files = finemap_chr22_ChipSeq_files,
  output_file = output_file_chr22_ChipSeq,
  panel = "HRC"
)

chr_21_chr_22_outer <- concat(list(ad_chr_21, ad_chr_22), join = "outer")
chr_22_ATAC_chr_22_ChipSeq_outer <- concat(list(ad_chr_22_ATAC, ad_chr_22_ChipSeq), join = "outer")
}
}
