% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.labf.R
\name{impute.labf}
\alias{impute.labf}
\title{Impute lABF Values for SNPs}
\usage{
impute.labf(snp, cred.set, cred.set.labf, imputed.labf)
}
\arguments{
\item{snp}{A character vector with names of SNPs.}

\item{cred.set}{A character vector of SNPs which are in the credible set.}

\item{cred.set.labf}{A numeric vector of lABF values for SNPs in the credible set.}

\item{imputed.labf}{A numeric value which will be used for imputation of lABF values for SNPs not in the credible set.}
}
\value{
A data frame with SNP names and their corresponding lABF values, either observed or imputed.
}
\description{
This function imputes lABF values for SNPs. For SNPs in the credible set, it assigns their corresponding lABF values.
For SNPs not in the credible set, it assigns an imputed lABF value.
}
\examples{
\dontrun{
snps <- c("rs1", "rs2", "rs3", "rs4")
credible_set <- c("rs1", "rs3")
credible_set_labf <- c(10.5, 12.3)
imputed_labf_value <- 5.0
impute.labf(snps, credible_set, credible_set_labf, imputed_labf_value)
}
}
