% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip.meta.R
\name{pip.meta}
\alias{pip.meta}
\title{Meta-analysis of PIPs for multiple credible sets}
\usage{
pip.meta(list_pips = NULL, verbose = FALSE)
}
\arguments{
\item{list_pips}{A SuSiE_R output (e.g. susieR::fitted_rss()).}

\item{verbose}{A boolean value. If TRUE, the log information will be printed.}
}
\value{
A single numveric vector of PIPs
}
\description{
This function inputs the list of vectors, each one containing PIPs for same
sets of SNPs and outputs a single vector with meta-analyzed PIPs.
}
\details{
This function inputs the list of vectors, each one containing PIPs for same
sets of SNPs and outputs a single vector with meta-analyzed PIPs.
}
\examples{
\dontrun{
pip.ma <- pip.meta(
    list_pips = susie_output,
    verbose = TRUE
)
}
}
