% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.cs.qc.R
\name{susie.cs.ht}
\alias{susie.cs.ht}
\title{Filters SuSiE output}
\usage{
susie.cs.ht(
  susie_output = NULL,
  pval = NULL,
  cs_lbf_thr = 2,
  signal_pval_threshold = 1,
  purity_mean_r2_threshold = 0.5,
  purity_min_r2_threshold = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{susie_output}{A SuSiE_R output (e.g. susieR::fitted_rss()).}

\item{pval}{A numeric vector representing P-values of the SNPs in the locus, tested in the susie_output.}

\item{cs_lbf_thr}{A numeric value representing the credible set logBF threshold for filtering credible sets (default is 2).}

\item{signal_pval_threshold}{A numeric value representing top SNP P-value threshold for filtering credible sets (default is 1e-6).}

\item{purity_mean_r2_threshold}{A numeric value representing the threshold for purity mean r2 QC metrics for filtering credible sets (default is 0.5).}

\item{purity_min_r2_threshold}{A numeric value representing the threshold for purity min r2 QC metrics for filtering credible sets (default is 0.5).}

\item{verbose}{A boolean value. If TRUE, the log information will be printed.}
}
\value{
A Susie object containing filtered fine-mapped credible sets.
}
\description{
This function filters the output from SuSiE and outputs Susie class object.
It return the object of class "susie" with filtered:
susie_output$sets$cs
susie_output$sets$purity
susie_output$sets$cs_index
susie_output$sets$coverage
}
\details{
This function takes the output from SuSiE, filters credible sets by purity metrics and CS lbf,
and returns a Susie object.
}
\examples{
\dontrun{
susie_filtered <- susie.cs.ht(
    susie_output = susie_output,
    pval = pval,
    cs_lbf_thr = 2,
    signal_pval_threshold = 1,
    purity_mean_r2_threshold = 0,
    purity_min_r2_threshold = 0,
    verbose = TRUE
)
}
}
