% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata2coloc.R
\name{anndata2coloc}
\alias{anndata2coloc}
\title{Perform Colocalization Analysis on AnnData Object}
\usage{
anndata2coloc(ad, coloc_input)
}
\arguments{
\item{ad}{An AnnData object containing genetic data.}

\item{coloc_input}{a path to a file with schema of pairwises test to run}
}
\value{
A list of colocalization results for each pair of traits.
}
\description{
This function performs colocalization analysis on an AnnData object. It computes the product of
the sparse matrix with its transpose to identify shared elements, retrieves trait information,
and performs colocalization testing on the identified pairs.
}
