% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata2coloc_input.R
\name{anndata2coloc_input}
\alias{anndata2coloc_input}
\title{Perform Colocalization Analysis on AnnData Object}
\usage{
anndata2coloc_input(ad)
}
\arguments{
\item{ad}{An AnnData object containing genetic data.}
}
\value{
A data.frame with colnames of t2, t1, t1_study_id,t1_phenotype_id, 
t1_top_pvalue, t2_study_id, t2_phenotype_id, t2_top_pvalue. Each row
represents a pariwise colocalization test which is needed to be completed. 
This aata.frame, written on hard drive as CSV file is an input for 
anndata2coloc function.
}
\description{
This function inputs AnnData objec and retrun csv file for coloc nextflow
pipeline. It computes the product of the sparse matrix with its transpose to
identify shared elements, retrieves trait information,#' and performs
colocalization testing on the identified pairs.
}
\examples{
\dontrun{
library(anndata)
library(data.table)
library(dplyr)

chr22_molQTL_ad <- read_h5ad("/group/pirastu/prj_013_horizontal_codec/2024_06_20_AnnData/HUVEC_chr22_combined_credible_sets.h5ad")

chr22_molQTL_ad$obs$study_id <- str_extract(chr22_molQTL_ad$obs$cs_name, "^[A-Za-z]+_chr[0-9]+")

chr22_molQTL_ad$obs$phenotype_id <- str_match(chr22_molQTL_ad$obs$cs_name, "chr[0-9]+_([^_]+_[0-9]+|ENSG[0-9]+)")[,2]

coloc_guide_table <- anndata2coloc_input(chr22_molQTL_ad)

fwrite(coloc_guide_table,file = "/Users/sodbo.sharapov/OneDrive - Htechnopole/00_Sodbo_Projects/anndata/data/coloc_duie_table.csv")
}

}
